requirements.txt
acres==0.3.0
altair==5.5.0
annotated-types==0.7.0
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
et_xmlfile==2.0.0
etelemetry==0.3.1
filelock==3.18.0
fitz==0.0.1.dev2
fsspec==2025.3.2
generativeai==0.0.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
html5lib==1.1
httplib2==0.22.0
huggingface-hub==0.30.1
idna==3.10
isodate==0.6.1
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
looseversion==1.3.0
lxml==5.3.1
MarkupSafe==3.0.2
narwhals==1.33.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.4
prov==2.0.1
puremagic==1.28
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pydeck==0.9.1
pydot==3.0.4
PyMuPDF==1.25.5
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
rsa==4.9
scipy==1.15.2
selenium==4.29.0
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
streamlit==1.44.1
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traits==7.0.2
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
upstox-python-sdk==2.14.0
uritemplate==4.1.1
urllib3==2.3.0
uuid==1.30
watchdog==6.0.0
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
XlsxWriter==3.2.2
